### 2018-12-24 #1142 @whedon
[PRE REVIEW]: ArviZ a unified library for exploratory analysis of Bayesian models in Python
**Submitting author:** @canyon289 (<a href="http://orcid.org/0000-0003-0501-6098">Ravin Kumar</a>)
**Repository:** <a href="https://github.com/arviz-devs/arviz" target ="_blank">https://github.com/arviz-devs/arviz</a>
**Version:** v.31
