### 2018-12-24 #1142 @arfon
@canyon289 - this PR should fix the paper metadata: https://github.com/arviz-devs/arviz/pull/460
