### 2018-12-24 be7d6b4
Update paper.md (#460)
