### 2018-12-24 #1142 @canyon289
@mattpitkin @lindemann09 @malmaud

Whedon recommended I find reviewers in this Google doc
https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822

Found rows listed both Bayesian Statistics and Python and was wondering if you'd like to review
