### 2018-12-25 #1142 @malmaud
I'd be happy to review this in the next few days.
