### 2018-12-26 db92018
Add additional documentation to compareplot
