### 2018-12-26 9232025
Add exception if waic or loo is not present
