### 2018-12-26 96d8f86
Add black formatting
