### 2018-12-26 f9f19a8
Fix linting errors
