### 2018-12-27 54dd7fe
add pystan3 extraction
