+++ arviz/data/io_pystan.py
@@ -23,7 +23,7 @@ class PyStanConverter:
         observed_data=None,
         log_likelihood=None,
         coords=None,
-        dims=None
+        dims=None,
     ):
         self.posterior = posterior
         self.posterior_predictive = posterior_predictive
@@ -33,6 +33,7 @@ class PyStanConverter:
         self.log_likelihood = log_likelihood
         self.coords = coords
         self.dims = dims
+
         import pystan
 
         self.pystan = pystan
@@ -63,26 +64,6 @@ class PyStanConverter:
     def sample_stats_to_xarray(self):
         """Extract sample_stats from posterior."""
         posterior = self.posterior
-        dtypes = {"divergent__": bool, "n_leapfrog__": np.int64, "treedepth__": np.int64}
-
-        ndraws = [s - w for s, w in zip(posterior.sim["n_save"], posterior.sim["warmup2"])]
-
-        extraction = OrderedDict()
-        for chain, (pyholder, ndraws) in enumerate(zip(posterior.sim["samples"], ndraws)):
-            if chain == 0:
-                for key in pyholder["sampler_param_names"]:
-                    extraction[key] = []
-            for key, values in zip(pyholder["sampler_param_names"], pyholder["sampler_params"]):
-                extraction[key].append(values[-ndraws:])
-
-        data = OrderedDict()
-        for key, values in extraction.items():
-            values = np.stack(values, axis=0)
-            dtype = dtypes.get(key)
-            values = values.astype(dtype)
-            name = re.sub("__$", "", key)
-            name = "diverging" if name == "divergent" else name
-            data[name] = values
 
         # copy dims and coords
         dims = deepcopy(self.dims) if self.dims is not None else {}
@@ -91,16 +72,12 @@ class PyStanConverter:
         # log_likelihood
         log_likelihood = self.log_likelihood
         if log_likelihood is not None:
-            log_likelihood_data = get_draws(posterior, variables=log_likelihood)
-            data["log_likelihood"] = log_likelihood_data[log_likelihood]
             if isinstance(log_likelihood, str) and log_likelihood in dims:
                 dims["log_likelihood"] = dims.pop(log_likelihood)
             if isinstance(log_likelihood, str) and log_likelihood in coords:
                 coords["log_likelihood"] = coords.pop(log_likelihood)
 
-        # lp__
-        stat_lp = get_draws(posterior, variables="lp__")
-        data["lp"] = stat_lp["lp__"]
+        data = get_sample_stats(posterior, log_likelihood)
 
         return dict_to_dataset(data, library=self.pystan, coords=coords, dims=dims)
 
@@ -133,31 +110,7 @@ class PyStanConverter:
     def sample_stats_prior_to_xarray(self):
         """Extract sample_stats_prior from prior."""
         prior = self.prior
-        dtypes = {"divergent__": bool, "n_leapfrog__": np.int64, "treedepth__": np.int64}
-
-        ndraws = [s - w for s, w in zip(prior.sim["n_save"], prior.sim["warmup2"])]
-
-        extraction = OrderedDict()
-        for chain, (pyholder, ndraws) in enumerate(zip(prior.sim["samples"], ndraws)):
-            if chain == 0:
-                for key in pyholder["sampler_param_names"]:
-                    extraction[key] = []
-            for key, values in zip(pyholder["sampler_param_names"], pyholder["sampler_params"]):
-                extraction[key].append(values[-ndraws:])
-
-        data = OrderedDict()
-        for key, values in extraction.items():
-            values = np.stack(values, axis=0)
-            dtype = dtypes.get(key)
-            values = values.astype(dtype)
-            name = re.sub("__$", "", key)
-            name = "diverging" if name == "divergent" else name
[…]
