### 2018-12-27 3b333c6
close old figures
