+++ examples/plot_kde.py
@@ -6,14 +6,22 @@ _thumb: .2, .8
 """
 import arviz as az
 import numpy as np
+import matplotlib.pyplot as plt
 
-az.style.use('arviz-darkgrid')
+# close all the figures, if open from previous commands
+plt.close("all")
 
-data = az.load_arviz_data('centered_eight')
+az.style.use("arviz-darkgrid")
+
+data = az.load_arviz_data("centered_eight")
 
 # Combine posterior draws for from xarray of (4,500) to ndarray (2000,)
 y_hat = np.concatenate(data.posterior_predictive["obs"].values)
 
-ax = az.plot_kde(y_hat, label='Estimated Effect\n of SAT Prep', rug=True,
-                 plot_kwargs={'linewidth': 2, 'color': 'black'},
-                 rug_kwargs={'color': 'black'})
+ax = az.plot_kde(
+    y_hat,
+    label="Estimated Effect\n of SAT Prep",
+    rug=True,
+    plot_kwargs={"linewidth": 2, "color": "black"},
+    rug_kwargs={"color": "black"},
+)
