+++ examples/plot_energy.py
@@ -5,8 +5,12 @@ Energy Plot
 _thumb: .7, .5
 """
 import arviz as az
+import matplotlib.pyplot as plt
 
-az.style.use('arviz-darkgrid')
+# close all the figures, if open from previous commands
+plt.close("all")
 
-data = az.load_arviz_data('centered_eight')
+az.style.use("arviz-darkgrid")
+
+data = az.load_arviz_data("centered_eight")
 az.plot_energy(data, figsize=(12, 8))
