### 2018-12-27 9e7b980
pyro < 0.3 doesn't support multiple chains (as far as I can tell)
