### 2018-12-27 d415113
remove function as per pull request comments
