### 2018-12-27 c6f2dec
run `black arviz/`
