### 2018-12-27 6e82307
1.Change the model_fn to a lamda 2.Add parameter-checking for model_fn and observed
