+++ arviz/tests/helpers.py
@@ -178,16 +178,13 @@ def tfp_noncentered_schools(data, draws, chains):
         ),
     )
 
-    def model_fn():
-        return schools_model(data["J"], data["sigma"].astype(np.float32))
-
     with tf.Session() as sess:
         [states_, _] = sess.run([states, kernel_results])
 
     data = from_tfp(
         states_,
         var_names=["mu", "tau", "theta_tilde"],
-        model_fn=model_fn,
+        model_fn=lambda: schools_model(data["J"], data["sigma"].astype(np.float32)),
         observed=data["y"].astype(np.float32),
     )
     return data
