### 2018-12-27 7ccd004
fix for from_pyro with multiple chains (#463)
* try fix

* pyro < 0.3 doesn't support multiple chains (as far as I can tell)

* remove function as per pull request comments

* run `black arviz/`
