+++ arviz/data/io_pyro.py
@@ -56,6 +56,8 @@ class PyroConverter:
         try:  # Try pyro>=0.3 release syntax
             data = {
                 name: np.expand_dims(samples.enumerate_support(), 0)
+                if self.posterior.num_chains == 1
+                else samples.enumerate_support()
                 for name, samples in self.posterior.marginal(
                     sites=self.latent_vars
                 ).empirical.items()
