### 2018-12-27 903be8c
don't pickle pystan3
