### 2018-12-27 a06263a
check __version__ for attr (#466)
LGTM
