### 2018-12-27 297eb81
Add docs and exception to plot compare (#461)
* Add additional documentation to compareplot

* Add exception if waic or loo is not present

* Add black formatting

* Fix linting errors
