### 2018-12-27 6ce0289
Update .travis.yml
