### 2018-12-27 b017c51
Use matrix in travis
