### 2018-12-27 7cb7a3f
Delete sampler from pyro posterior (enable pickle)
