+++ arviz/tests/helpers.py
@@ -133,6 +133,7 @@ def pyro_centered_schools(data, draws, chains):
     posterior.kernel = None
     posterior.run = None
     posterior.logger = None
+    posterior.sampler = None
     return posterior
 
 
