### 2018-12-27 d3088c4
squeeze 0.3 output
