### 2018-12-27 58ce4c6
replace newline marker with _
