### 2018-12-28 4a3ce6e
WIP: Adding better error checking to the build script
