### 2018-12-28 693d62f
check if pyro posterior has sampler and squeeze observed data
