+++ .travis.yml
@@ -7,6 +7,9 @@ matrix:
   - name: "Python 3.6 Unit Test"
     python: 3.6
     env: PYSTAN_VERSION=latest NAME="UNIT"
+  - name: "Python 3.6 Unit Test - PyStan preview"
+    python: 3.6
+    env: PYSTAN_VERSION=preview NAME="UNIT"
   - name: "Python 3.5 Unit Test"
     python: 3.5
     env: PYSTAN_VERSION=latest NAME="UNIT"
