### 2018-12-28 8f22546
fix pystan tests
