### 2018-12-28 f8dd22a
fix pystan tests
