### 2018-12-28 0c93295
fix dtype
