### 2018-12-28 d6eb4fa
Add it to docker instructions (#471)
