### 2018-12-28 aabf6eb
enable densityplot error
