+++ arviz/tests/test_data.py
@@ -262,7 +262,7 @@ class TestDictNetCDFUtils:
         return Data
 
     def check_var_names_coords_dims(self, dataset):
-        assert set(dataset.data_vars) == {"mu", "tau", "theta_tilde", "theta"}
+        assert set(dataset.data_vars) == {"mu", "tau", "eta", "theta"}
         assert set(dataset.coords) == {"chain", "draw", "school"}
 
     def get_inference_data(self, data, eight_schools_params):
@@ -270,7 +270,7 @@ class TestDictNetCDFUtils:
             data.obj,
             group="posterior",
             coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"theta": ["school"], "theta_tilde": ["school"]},
+            dims={"theta": ["school"], "eta": ["school"]},
         )
 
     def test_convert_to_inference_data(self, data, eight_schools_params):
@@ -283,7 +283,7 @@ class TestDictNetCDFUtils:
             data.obj,
             group="posterior",
             coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"theta": ["school"], "theta_tilde": ["school"]},
+            dims={"theta": ["school"], "eta": ["school"]},
         )
         assert dataset.draw.shape == (draws,)
         assert dataset.chain.shape == (chains,)
@@ -328,7 +328,7 @@ class TestPyMC3NetCDFUtils:
             prior=prior,
             posterior_predictive=posterior_predictive,
             coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"theta": ["school"], "theta_tilde": ["school"]},
+            dims={"theta": ["school"], "eta": ["school"]},
         )
 
     def test_sampler_stats(self, data, eight_schools_params):
@@ -369,7 +369,7 @@ class TestPyStanNetCDFUtils:
                 "y": ["school"],
                 "log_lik": ["school"],
                 "y_hat": ["school"],
-                "theta_tilde": ["school"],
+                "eta": ["school"],
             },
         )
 
@@ -390,7 +390,7 @@ class TestPyStanNetCDFUtils:
                 "theta": ["school"],
                 "y": ["school"],
                 "y_hat": ["school"],
-                "theta_tilde": ["school"],
+                "eta": ["school"],
                 "log_lik": ["log_likelihood_dim"],
             },
         )
@@ -404,12 +404,7 @@ class TestPyStanNetCDFUtils:
             prior_predictive=["y_hat", "log_lik"],
             observed_data="y",
             coords={"school": np.arange(eight_schools_params["J"])},
-            dims={
-                "theta": ["school"],
-                "y": ["school"],
-                "y_hat": ["school"],
-                "theta_tilde": ["school"],
-            },
+            dims={"theta": ["school"], "y": ["school"], "y_hat": ["school"], "eta": ["school"]},
         )
 
     def get_inference_data4(self, data):
@@ -594,7 +589,7 @@ class TestCmdStanNetCDFUtils:
                     "y": ["school"],
                     "log_lik": ["school"],
                     "y_hat": ["school"],
-                    "theta_tilde": ["school"],
+                    "eta": ["school"],
                 },
             )
             assert hasattr(inference_data, "posterior")
@@ -626,7 +621,7 @@ class TestCmdStanNetCDFUtils:
                     "y": ["school"],
                     "log_lik": ["school"],
                     "y_hat": ["school"],
-                    "theta_tilde": ["school"],
+                    "eta": ["school"],
                 },
             )
             assert hasattr(inference_data, "posterior")
@@ -655,12 +650,7 @@ class TestCmdStanNetCDFUtils:
                 observed_data_var=["y"],
                 log_likelihood="log_lik",
                 coords={"school": np.arange(8), "log_lik_dim_0": np.arange(8)},
[…]
