### 2018-12-29 aa8ba21
Update .travis.yml (#467)
* Update .travis.yml

* Use matrix in travis

* Delete sampler from pyro posterior (enable pickle)

* squeeze 0.3 output

* replace newline marker with _

* check if pyro posterior has sampler and squeeze observed data

* Undo unneeded fix
