### 2018-12-29 2892881
Add Docker Testing to travisCI
