### 2018-12-29 35f22bb
Fix docker commands
