### 2018-12-29 21b1337
Fix linting docker command
