### 2018-12-29 22953bf
Fix sphinx and linting
