### 2018-12-29 09bc4d6
Add env variables for sphinx and pandoc to container
