+++ .travis.yml
@@ -52,13 +52,15 @@ script:
   # - if [[ $NAME == SPHINX ]]; then travis-sphinx build --nowarn --source=doc ; fi
 
   - if [[ $NAME == SPHINX ]]; then docker run --name arviz_sphinx arviz:latest /bin/bash -c \
-                              "source /root/.bashrc && travis-sphinx build --nowarn --source=doc";
+                              "source /root/.bashrc && \
+                               export LC_ALL=C.UTF-8 && \
+                               export LANG=C.UTF-8 && \
+                               travis-sphinx build --nowarn --source=doc"
     fi
 after_success:
   # - if [[ $NAME == SPHINX ]]; then travis-sphinx deploy; fi
-  - if [[ $NAME == SPHINX ]]; then docker exec /bin/bash -c \
-                              "source /root/.bashrc
-                               && export LC_ALL=C.UTF-8
-                               &&export LANG=C.UTF-8
-                               && travis-sphinx deploy"; fi
-  - if [[ $NAME == UNIT ]]; then coveralls ; fi
+  - if [[ $NAME == SPHINX ]]; then docker exec arviz_sphinx /bin/bash -c \
+                               travis-sphinx deploy"; fi
+  # - if [[ $NAME == UNIT ]]; then coveralls ; fi
+  - if [[ $NAME == UNIT ]]; then docker run arviz:latest /bin/bash -c \
+                                "source /root/.bashrc && coveralls" ; fi
