### 2018-12-29 412304c
Fix closing quotation issue
