+++ .travis.yml
@@ -63,7 +63,7 @@ script:
 after_success:
   # - if [[ $NAME == SPHINX ]]; then travis-sphinx deploy; fi
   - if [[ $NAME == SPHINX ]]; then docker exec arviz_sphinx /bin/bash -c \
-                               travis-sphinx deploy"; fi
+                               "travis-sphinx deploy"; fi
   # - if [[ $NAME == UNIT ]]; then coveralls ; fi
   - if [[ $NAME == UNIT ]]; then docker run arviz:latest /bin/bash -c \
                                 "source /root/.bashrc && coveralls" ; fi
