### 2018-12-29 b77b93d
Add coverage to default docker testing
