### 2018-12-29 ec5c5b6
Add environmental vars in docker run command
