+++ arviz/tests/test_data.py
@@ -23,6 +23,7 @@ from .helpers import (  # pylint: disable=unused-import
     load_cached_models,
     pystan_extract_unpermuted,
     stan_extract_dict,
+    pystan_version,
 )
 
 
@@ -246,13 +247,13 @@ class TestDictNetCDFUtils:
 
         class Data:
             _, stan_fit = load_cached_models(eight_schools_params, draws, chains)["pystan"]
-            try:
+            if pystan_version() == 2:
                 stan_dict = pystan_extract_unpermuted(stan_fit)
                 obj = {}
                 for name, vals in stan_dict.items():
                     if name not in {"y_hat", "log_lik"}:  # extra vars
                         obj[name] = np.swapaxes(vals, 0, 1)
-            except:
+            else:
                 stan_dict = stan_extract_dict(stan_fit)
                 obj = {}
                 for name, vals in stan_dict.items():
