### 2018-12-29 031ad59
fix docstring and remove pylint comments
