### 2018-12-29 978d682
Remove warmup from pystan
