### 2018-12-29 31ff663
Print docker version
