### 2018-12-29 89f86e7
Revert travis sphinx build
