+++ .travis.yml
@@ -55,12 +55,11 @@ script:
 
   # - if [[ $NAME == SPHINX ]]; then travis-sphinx build --nowarn --source=doc ; fi
 
-  - if [[ $NAME == SPHINX ]]; then docker run --name arviz_sphinx \
-                              -e LC_ALL="C.UTF-8" \
-                              -e LANG="C.UTF-8" \
-                              arviz:latest /bin/bash -c \
-                              "source /root/.bashrc && \
-                               travis-sphinx build --nowarn --source=doc";
+  - if [[ $NAME == SPHINX ]]; then docker run --name arviz_sphinx arviz:latest /bin/bash -c \
+                                              "source /root/.bashrc && \
+                                              export LC_ALL=C.UTF-8 && \
+                                              export LANG=C.UTF-8 && \
+                                              travis-sphinx build --nowarn --source=doc"
     fi
 
 after_success:
