### 2018-12-29 78c4c8d
Another try at travis-sphinx
