### 2018-12-29 cba3545
Add language env variables to dockerfile
