+++ .travis.yml
@@ -57,9 +57,7 @@ script:
 
   - if [[ $NAME == SPHINX ]]; then docker run --name arviz_sphinx arviz:latest /bin/bash -c \
                                               "source /root/.bashrc && \
-                                              export LC_ALL=C.UTF-8 && \
-                                              export LANG=C.UTF-8 && \
-                                              travis-sphinx build --nowarn --source=doc";
+                                               travis-sphinx build --nowarn --source=doc";
     fi
 
 after_success:
