### 2018-12-29 ff28dcb
Remove latest image tag
