### 2018-12-29 5e0fafc
Add more debugging info to travis file
