### 2018-12-29 29da59b
Fix doc deploy command
