### 2018-12-29 8ce0af1
Add coveralls install
