### 2018-12-29 7b2578d
Fix spaces in travis-sphinx deploy
