### 2018-12-29 17fa7af
Add travis branch env var to arviz_sphinx container
