### 2018-12-29 efe7c6b
Add env vars to coveralls and travis-sphinx
