### 2018-12-29 ff8b886
Fix placement of env variable for travis-sphinx
