### 2018-12-29 6b0e0a3
Fix env variable arguments yet again
