### 2018-12-29 20329e3
Add more env variables and coveralls token
