### 2018-12-29 e8b32d9
Add more env vars into travis-sphinx
