### 2018-12-29 4feb8ac
Attempt to fix line continuation
