### 2018-12-29 9d19a3b
Export coveralls env var and single line travis-sphinx
