### 2018-12-29 eb7c7e9
hard code data for pystan3
