### 2018-12-29 12f2ccb
add models for pystan tests
