### 2018-12-30 4b2098b
Add comments and remove unneeded code
