### 2018-12-30 f68296f
Make testing more explicit in travis file
