### 2018-12-30 588658f
Try travis file copy hack for coveralls
