+++ .travis.yml
@@ -67,5 +67,6 @@ script:
     fi
 
 after_success:
+  - ls -la
   - if [[ $NAME == SPHINX ]]; then travis-sphinx deploy; fi
-  - if [[ $NAME == UNIT ]]; then coveralls ; fi
+  - if [[ $NAME == UNIT ]]; then cp ../arviz /opt/arviz && coveralls ; fi
