### 2018-12-30 d3d2c04
Add .coveragerc file
