### 2018-12-30 aad23df
Add sudo to cp
