+++ .travis.yml
@@ -70,4 +70,4 @@ after_success:
   - echo $(pwd)
   - ls -la
   - if [[ $NAME == SPHINX ]]; then travis-sphinx deploy; fi
-  - if [[ $NAME == UNIT ]]; then sudo cp -r ../arviz /opt && coveralls coveralls; fi
+  - if [[ $NAME == UNIT ]]; then sudo cp -r ../arviz /opt && coveralls; fi
