### 2018-12-30 aa3b1dd
[WIP] Add Docker Testing to travisCI (#473)
* Add Docker Testing to travisCI

* Fix docker commands

* Fix linting docker command

* Fix sphinx and linting

* Fix sphinx build error

* Add env variables for sphinx and pandoc to container

* Fix travis file

* Fix closing quotation issue

* Add coverage to default docker testing

* Fix unexpected end of file problem

* Add environmental vars in docker run command

* Update docker version

* Print docker version

* Another try with sphinx command

* Revert travis sphinx build

* Another try at travis-sphinx

* Practice makes perfect

* Add langu
[…]
