### 2018-12-30 67748aa
Instructions for PyStan3 model object
