### 2018-12-30 6b7b98e
change .coverage paths manually with sed
