### 2018-12-30 0f6ca94
Support PyStan3 (#464)
* add pystan3 extraction

* fix tests

* don't pickle pystan3

* pystan preview

* fix pystan tests

* theta_tilde to eta

* test pystan version

* fix docstring and remove pylint comments

* Fix typo

* Remove warmup from pystan

* skip pre-releases

* hard code data for pystan3

* add models for pystan tests

* Instructions for PyStan3 model object

* fix whitespace
