### 2018-12-30 972b79b
test with pyro=0.2.1
