+++ scripts/Dockerfile
@@ -7,9 +7,11 @@ ARG SRC_DIR
 
 ARG PYTHON_VERSION
 ARG PYSTAN_VERSION
+ARG PYRO_VERSION
 
 ENV PYTHON_VERSION=${PYTHON_VERSION}
 ENV PYSTAN_VERSION=${PYSTAN_VERSION}
+ENV PYRO_VERSION=${PYRO_VERSION}
 ENV DOCKER_BUILD=true
 
 # For Sphinx documentation builds
