### 2018-12-31 450af16
test with pyro==0.2.1 (#475)
* test with pyro=0.2.1

* name for env
