### 2018-12-31 3bdd983
Fix linting errors
