### 2018-12-31 2b63458
fix ridge input
