### 2018-12-31 fe29735
fix line-too-long and return ax from forestplot
