### 2018-12-31 eb4f1e2
Fix pylint false-positive error
