### 2018-12-31 f1fd5d1
split rhat
