### 2018-12-31 08715ff
Move pylint comment
