+++ arviz/plots/forestplot.py
@@ -192,7 +192,9 @@ def plot_forest(
 
 class PlotHandler:
     """Class to handle logic from ForestPlot."""
-
+    
+    # pylint: disable=inconsistent-return-statements
+    
     def __init__(self, data, var_names, model_names, combined, colors):
         if not isinstance(data, (list, tuple)):
             data = [data]
@@ -350,7 +352,6 @@ class PlotHandler:
         elif len(rope) == 2:
             ax.axvspan(rope[0], rope[1], 0, self.y_max(), color="C2", alpha=0.5)
         else:
-            # pylint: disable=inconsistent-return-statements
             raise ValueError(
                 "Argument `rope` must be None, a dictionary like"
                 '{"var_name": {"rope": (lo, hi)}}, or an '
