### 2018-12-31 3bd185c
fix rhat docstring and test for ndarray
