### 2018-12-31 b865cd4
Changing travis.yml and removing b
