### 2018-12-31 f79ffe8
Add some more information to the inference data of tfp (#447)
* Add some more information to the InferenceData from tfp, including:
1. observed data
2. posterior predictive
3. log-likelihood of observed data

* correct the format

* 1.Change the model_fn to a lamda
2.Add parameter-checking for model_fn and observed
