### 2018-12-31 7f189c5
Use Split R-hat (#477)
* rename rhats

* split rhat

* fix forestplot

* fix rhats

* fix tests

* fix missing 2

* fix rhat docstring and test for ndarray
