### 2018-12-31 9509e25
display quartiles
