### 2018-12-31 6669d0c
Fix CoC link in README
