### 2018-12-31 arviz-devs/arviz#481 @malmaud
README doesn't mention non-master installation instructions.
In the README installation section, you don't mention `pip install arviz` as the way of installing the release version. I know it seems trivial, but some people might not realize if the instructions aren't explicit.
