### 2018-12-31 arviz-devs/arviz#481 @ahartikainen
Oh, we have forgot to update readme.

Do you want to do a PR?
