### 2018-12-31 #1143 @malmaud
OK, I've played with all the functionality of the software on my own device - everything works well, the utility is clear, and the documentation is comprehensive (albeit with a few minor typos). I think this is an easy accept after just a few tiny fixes:

* Fix https://github.com/arviz-devs/arviz/issues/481
* In the paper, Rabin's affiliation is listed as 'None'. Is that really true, or is there a workplace you could list?
* Period missing after '...analyzes"', and I think you want 'analyses' there.
* A figure showing a gallery of plot types would be a useful addition to the paper.
* Figure 1 isn't referenced in the text of the paper and its caption isn't particularly informative.  The paragraph that talks about soring datasets on disks should reference that figure and its caption should be made informative.
