### 2019-01-01 arviz-devs/arviz#481 @canyon289
Definitely not trivial. This is a good suggestion.
