### 2019-01-01 48cd609
Add more installation instructions to README (#482)
