### 2019-01-01 arviz-devs/arviz#481 @canyon289
Closed by https://github.com/arviz-devs/arviz/pull/482
