### 2019-01-01 812d81f
Fix analyses typo and add period
