+++ paper/paper.md
@@ -43,7 +43,7 @@ When working with Bayesian models there are a series of related tasks that need 
 - Comparison of models, including model selection or model averaging
 - Preparation of the results for a particular audience
 
-Successfully performing such tasks are central to the iterative and interactive modeling process. These tasks require both numerical and visual summaries to help statisticians or practitioners analyze visual summaries. In the words of Persi Diaconis [@DiaconisTheoriesDataAnalysis2011] "Exploratory data analysis seeks to reveal structure, or simple descriptions in data. We look at numbers or graphs and try to find patterns. We pursue leads suggested by background information, imagination, patterns perceived, and experience with other data analyzes"
+Successfully performing such tasks are central to the iterative and interactive modeling process. These tasks require both numerical and visual summaries to help statisticians or practitioners analyze visual summaries. In the words of Persi Diaconis [@DiaconisTheoriesDataAnalysis2011] "Exploratory data analysis seeks to reveal structure, or simple descriptions in data. We look at numbers or graphs and try to find patterns. We pursue leads suggested by background information, imagination, patterns perceived, and experience with other data analyses".
 
 For these reasons we introduce ArviZ, a Python package for exploratory analysis of Bayesian models. ArviZ aims to be a package that integrates seamlessly with established probabilistic programming languages like PyStan [@StanProbabilisticProgramming], PyMC [@SalvatierProbabilisticProgrammingPython2016], Edward [@TranDeepProbabilisticProgramming2017; @TranEdwardlibraryprobabilistic2016], emcee [@emcee], Pyro [@bingham2018pyro], and easily integrated with novel or bespoke Bayesian analyses.  Where the aim of the probabilistic programming languages is to make it easy to build and solve Bayesian models, the aim of the ArviZ library is to make it easy to process and analyze the results from the Bayesian models. We hope ArviZ will become a key Python tool for Bayesian data analysis by allowing users to focus on problems from their domain knowledge and not on computational details.
 
