### 2019-01-01 99cc4e7
Add reference to figure and update caption
Put more linebreaks to make editing easier
