### 2019-01-01 #1143 @canyon289
@malmaud

Thanks for the feedback. A PR is open, just waiting for feedback internally before we submit it again for review to you.

https://github.com/arviz-devs/arviz/pull/483
