### 2019-01-01 8009401
Slightly reword Contributions section
