### 2019-01-02 0f54b8e
Additional edits to paper (#483)
* Fix analyses typo and add period

* Add employment and country

* Add reference to figure and update caption

Put more linebreaks to make editing easier

* Add example gallery

* Slightly reword Contributions section
