### 2019-01-02 #1143 @canyon289
@malmaud The content is ready to review but the figure gallery seems to have rendered oddly. Is there a way I can see what command whedon is running to compile the pdf so I can test locally?
