### 2019-01-02 #1143 @mattpitkin
@canyon289 you can see a Makefile with an example of how you can locally generate a LaTeX file and pdf from the markdown doc [here](https://github.com/mattpitkin/psrqpy/blob/master/paper/Makefile).
