### 2019-01-02 045f348
compute quantiles from kde
