### 2019-01-02 d95ec1d
remove unused variable
