### 2019-01-02 5e1322c
Fix example gallery
