### 2019-01-02 4d5d631
Fix example gallery in paper (#484)
* Fix example gallery

* Add revision with whitespace hack

Header order preserved
