### 2019-01-02 #1143 @canyon289
Tried getting the PDF to render correctly I think we might have figured it out

PR for reference
https://github.com/arviz-devs/arviz/pull/484

Here it goes
