### 2019-01-02 d07677e
Reduce paper to 2 figures
