### 2019-01-02 714bc2d
Paper fix 2 (#485)
* Reduce paper to 2 figures

* Edit verbiage slightly
