### 2019-01-02 #1143 @canyon289
Still not correct. I will revisit the way I'm rendering these to see if I can test more effectively in my local environment.
