### 2019-01-03 #1143 @canyon289
@malmaud and @mattpitkin

All feedback has been addressed, except example gallery which I'm struggling with tremendously.

I'm having issues generating a PDF with the make file, my tex output displays errors.

I can get a PDF to render with the command
```
pandoc --filter pandoc-citeproc --bibliography=paper.bib  --variable papersize=a4paper -s paper.md -o paper.pdf
```
But while the output of this command may look good, the whedon generated pdf does not.

Lastly I tried looking for other JOSS papers with two or more figures but my random search was fruitless.

Here are two versions of the paper with Example gallery. Any suggestions on how to get them to render nicely? I realize this may be outside of the normal review role, I apologize for the trouble.

Multiple Figures
https://github.com/arviz-devs/arviz/commit/4d5d6319c8b80872292ab2589606ca6d4a9b183a

Two Figures
https://github.com/arviz-devs/arviz/commit/714bc2d3d0b852e477ef87c6460c3413411aa128
