### 2019-01-03 #1143 @arfon
@canyon289 - could you share the PDF you've compiled locally here? I'd like to bette understand what you're trying to accomplish.
