### 2019-01-03 #1143 @canyon289
We're trying to add a local gallery per the [review feedback](https://github.com/openjournals/joss-reviews/issues/1143#issuecomment-450685429
) In other words we'd like to meet the intent of the review, but struggling to get nice formatting while doing so.

Here's the two approaches we tired.

**2+ Figures**
Here's a local render with multiple figures. Unfortunately this render required a "hack" to get it to work locally, so not a great solution
https://github.com/arviz-devs/arviz/files/2722319/paper.pdf

Header hack
https://github.com/arviz-devs/arviz/pull/484/files#diff-e2778e7674f45b806282a9611faa7220R100

**Two Figures**
We also tried a version with just two figures. In the local render it seems to work, but not in JOSS.
https://github.com/arviz-devs/arviz/files/2722350/paper.pdf
https://github.com/arviz-devs/arviz/pull/485/files

Let me know if I explained this poorly and I can clarify any details
