### 2019-01-03 5ca831f
normalize input
