+++ arviz/data/io_emcee.py
@@ -94,7 +94,7 @@ class EmceeConverter:
         )
 
 
-def from_emcee(sampler, *, var_names=None, arg_names=None, coords=None, dims=None):
+def from_emcee(sampler=None, *, var_names=None, arg_names=None, coords=None, dims=None):
     """Convert emcee data into an InferenceData object.
 
     Parameters
@@ -111,5 +111,5 @@ def from_emcee(sampler, *, var_names=None, arg_names=None, coords=None, dims=Non
         Map variable names to their coordinates
     """
     return EmceeConverter(
-        sampler, var_names=var_names, arg_names=arg_names, coords=coords, dims=dims
+        sampler=sampler, var_names=var_names, arg_names=arg_names, coords=coords, dims=dims
     ).to_inference_data()
