### 2019-01-03 ef3fc5b
from_pyro and from_tfp
