### 2019-01-03 #1143 @arfon
Ah OK. The problem here is that Pandoc + Markdown offers very little in terms of document flow/layout control.

Based on version https://github.com/arviz-devs/arviz/commit/4d5d6319c8b80872292ab2589606ca6d4a9b183a, I've generated the PDF locally by moving the `# Example plots` section _below_ the `# Acknowledgments` section and slightly tweaking the width of the figures i.e.:

```
# Example plots

A portion of ArviZ's functionality is shown in Figure 2 through Figure 5.
ArviZ supports more plotting and inference diagnostics, in addition to the ones shown here. Additional examples can be found ArviZ documentation gallery

![Bivariate hexbin plot with marginal distributions](plot_joint.png){width=80%}

![2D Kernel Density estimation](plot_kde_2d.png){width=80%}

![Markov Chain Monte Carlo Trace Plot](plot_trace.png){width=80%}

![John Kruschke styled posterior distribution plots](plot_posterior.png){width=80%}
```

This gives the resulting output: [10.21105.joss.01143.pdf](https://github.com/openjournals/joss-reviews/files/2723702/10.21105.joss.01143.pdf)

What do you think?
