### 2019-01-03 cd81e35
fix tfp test output
