### 2019-01-03 02eca04
remove impossible lines, test some edge cases
