### 2019-01-04 e11432b
Change gallery position in JOSS paper (#493)
Reword Example Plots section slightly
