### 2019-01-04 #1143 @canyon289
Looks good. We made the changes in our paper. Hopefully this works
