### 2019-01-04 #1143 @canyon289
This looks ready for rereview. Thank you all for the help
