### 2019-01-04 #1143 @mattpitkin
Hi @canyon289, everything looks good, but I just have a few comments:

* the API documentation appears complete for the methods, but doesn't have any class documentation. In particular, I think it would be useful to have the documentation for the `InferenceData` class, with that containing some examples or descriptions of what the class can store.
* for the _Pyro: Deep Universal Probabilistic Programming_ reference could you add the arXiv link https://arxiv.org/abs/1810.09538 as it doesn't seem to be released in JMLR yet.
* for the _A Widely Applicable Bayesian Information Criterion_ maybe also add a link to the arXiv version of the paper https://arxiv.org/abs/1208.6338, or the JMLR webpage http://www.jmlr.org/papers/v14/watanabe13a.html
