### 2019-01-04 a139a76
deprecate load_data and save_data
