+++ arviz/tests/helpers.py
@@ -128,9 +128,9 @@ def pyro_centered_schools(data, draws, chains):
     sigma = torch.Tensor(data["sigma"]).type(torch.Tensor)
 
     nuts_kernel = NUTS(_pyro_conditioned_model, adapt_step_size=True)
-    posterior = MCMC(  # pylint:disable=not-callable
-        nuts_kernel, num_samples=draws, warmup_steps=500
-    ).run(_pyro_centered_model, sigma, y)
+    posterior = MCMC(nuts_kernel, num_samples=draws, warmup_steps=500).run(
+        _pyro_centered_model, sigma, y
+    )  # pylint:disable=not-callable
 
     # This block lets the posterior be pickled
     for trace in posterior.exec_traces:
