### 2019-01-04 24cad9f
Increase PyStan testing (#494)
* remove impossible lines, test some edge cases

* fix typo
