### 2019-01-05 e7a7742
Normalize from_xyz functions (#490)
* normalize input

* fix tests

* fix kw

* import error

* linting issues

* Missing s

* from_pyro and from_tfp

* fix tfp test output

* deprecate load_data and save_data

* run black

* fix black
