+++ arviz/tests/helpers.py
@@ -350,11 +350,7 @@ def stan_extract_dict(fit, var_names=None):
             continue
 
         # in future fix the correct number of draws if fit.save_warmup is True
-        new_shape = (
-            *fit.dims[fit.param_names.index(var)],
-            -1,
-            fit.num_chains,
-        )  # pylint: disable=protected-access
+        new_shape = (*fit.dims[fit.param_names.index(var)], -1, fit.num_chains)
         values = fit._draws[fit._parameter_indexes(var), :]  # pylint: disable=protected-access
         values = values.reshape(new_shape, order="F")
         values = np.moveaxis(values, [-2, -1], [1, 0])
