### 2019-01-05 cd6afe1
add tests and fix bugs
