+++ arviz/tests/test_stats.py
@@ -1,12 +1,13 @@
 # pylint: disable=redefined-outer-name
-
+from copy import deepcopy
 import numpy as np
 from numpy.testing import assert_almost_equal, assert_array_almost_equal, assert_array_less
 import pytest
 from scipy.stats import linregress
 
 from ..data import load_arviz_data
-from ..stats import bfmi, compare, hpd, r2_score, waic, psislw, summary
+from ..stats import bfmi, compare, hpd, loo, r2_score, waic, psislw, summary
+from ..stats.stats import _gpinv, _mc_error
 
 
 @pytest.fixture(scope="session")
@@ -32,11 +33,26 @@ def test_hpd():
     assert_array_almost_equal(interval, [-1.88, 1.88], 2)
 
 
+def test_hpd_bad_ci():
+    normal_sample = np.random.randn(10)
+    with pytest.raises(ValueError):
+        hpd(normal_sample, credible_interval=2)
+
+
 def test_r2_score():
     x = np.linspace(0, 1, 100)
     y = np.random.normal(x, 1)
     res = linregress(x, y)
     assert_almost_equal(res.rvalue ** 2, r2_score(y, res.intercept + res.slope * x).r2, 2)
+
+
+def test_r2_score_multivariate():
+    x = np.linspace(0, 1, 100)
+    y = np.random.normal(x, 1)
+    res = linregress(x, y)
+    y_multivariate = np.c_[y, y]
+    y_multivariate_pred = np.c_[res.intercept + res.slope * x, res.intercept + res.slope * x]
+    assert not np.isnan(r2_score(y_multivariate, y_multivariate_pred).r2)
 
 
 @pytest.mark.parametrize("method", ["stacking", "BB-pseudo-BMA", "pseudo-BMA"])
@@ -48,6 +64,14 @@ def test_compare_same(centered_eight, method):
     assert_almost_equal(np.sum(weight), 1.0)
 
 
+def test_compare_unknown_ic_and_method(centered_eight, non_centered_eight):
+    model_dict = {"centered": centered_eight, "non_centered": non_centered_eight}
+    with pytest.raises(NotImplementedError):
+        compare(model_dict, ic="Unknown", method="stacking")
+    with pytest.raises(ValueError):
+        compare(model_dict, ic="loo", method="Unknown")
+
+
 @pytest.mark.parametrize("ic", ["waic", "loo"])
 @pytest.mark.parametrize("method", ["stacking", "BB-pseudo-BMA", "pseudo-BMA"])
 def test_compare_different(centered_eight, non_centered_eight, ic, method):
@@ -55,6 +79,20 @@ def test_compare_different(centered_eight, non_centered_eight, ic, method):
     weight = compare(model_dict, ic=ic, method=method)["weight"]
     assert weight["non_centered"] > weight["centered"]
     assert_almost_equal(np.sum(weight), 1.0)
+
+
+def test_compare_different_size(centered_eight, non_centered_eight):
+    centered_eight = deepcopy(centered_eight)
+    centered_eight.posterior = centered_eight.posterior.drop("Choate", "school")
+    centered_eight.sample_stats = centered_eight.sample_stats.drop("Choate", "school")
+    centered_eight.posterior_predictive = centered_eight.posterior_predictive.drop(
+        "Choate", "school"
+    )
+    centered_eight.prior = centered_eight.prior.drop("Choate", "school")
+    centered_eight.observed_data = centered_eight.observed_data.drop("Choate", "school")
+    model_dict = {"centered": centered_eight, "non_centered": non_centered_eight}
+    with pytest.raises(ValueError):
+        compare(model_dict, ic="waic", method="stacking")
 
 
[…]
