+++ arviz/tests/test_diagnostics.py
@@ -5,6 +5,8 @@ import pytest
 
 from ..data import load_arviz_data
 from ..stats import rhat, effective_sample_size, geweke
+from ..stats.diagnostics import ks_summary
+
 
 GOOD_RHAT = 1.1
 
@@ -33,10 +35,22 @@ class TestDiagnostics:
         r_hat = rhat(np.vstack([20 + np.random.randn(1, 100), np.random.randn(1, 100)]))
         assert 1 / GOOD_RHAT > r_hat or GOOD_RHAT < r_hat
 
+    def test_rhat_bad_shape(self):
+        with pytest.raises(TypeError):
+            rhat(np.random.randn(3))
+
     def test_effective_sample_size_array(self):
         eff_n_hat = effective_sample_size(np.random.randn(4, 100))
         assert eff_n_hat > 100
         assert eff_n_hat < 800
+
+    def test_effective_sample_size_bad_shape(self):
+        with pytest.raises(TypeError):
+            effective_sample_size(np.random.randn(3))
+
+    def test_effective_sample_size_bad_chains(self):
+        with pytest.raises(TypeError):
+            effective_sample_size(np.random.randn(1, 3))
 
     @pytest.mark.parametrize("var_names", (None, "mu", ["mu", "tau"]))
     def test_effective_sample_size_dataset(self, data, var_names):
@@ -56,3 +70,25 @@ class TestDiagnostics:
 
         assert gw_stat.shape[0] == intervals
         assert 10000 * last - gw_stat[:, 0].max() == 1
+
+    def test_geweke_bad_interval(self):
+        # lower bound
+        with pytest.raises(ValueError):
+            geweke(np.random.randn(10), first=0)
+        # upper bound
+        with pytest.raises(ValueError):
+            geweke(np.random.randn(10), last=1)
+        # sum larger than 1
+        with pytest.raises(ValueError):
+            geweke(np.random.randn(10), first=0.9, last=0.9)
+
+    def test_ks_summary(self):
+        """Instead of psislw data, this test uses fake data."""
+        pareto_tail_indices = np.array([0.1, 0.1, 0.1, 0.2, 0.2, 0.2, 0.2])
+        with pytest.warns(UserWarning):
+            summary = ks_summary(pareto_tail_indices)
+        assert summary is not None
+        pareto_tail_indices2 = np.array([0.1, 0.1, 0.1, 0.2, 0.2, 0.2, 0.6])
+        with pytest.warns(UserWarning):
+            summary2 = ks_summary(pareto_tail_indices2)
+        assert summary2 is not None
