### 2019-01-06 19a13ed
Rename n_eff to s_eff
