### 2019-01-06 0684ec8
Rename tests for s_eff
