### 2019-01-07 #1143 @mattpitkin
@canyon289 thanks for the changes. Would it be possible to include the `InferenceData` class in the API documentation, or is there a specific reason for not doing this? I think it would be good to have it included even if the docstring mainly just adds a link to the page https://arviz-devs.github.io/arviz/notebooks/XarrayforArviZ.html. Also, could that notebook also contain the plot from Figure 1 of the paper?
