### 2019-01-07 8b9528a
Remove html for proper github rendering
