### 2019-01-07 #1143 @canyon289
@mattpitkin
My (subjective) reasoning for not including `az.InferenceData` in the api docs was because all the methods in the api docs are intended to be callable by the user. For InferenceData the users don't directly instantiate an `az.InferenceData`. This is why I after some debate I opted not to include it in api docs, however this reasoning is largely subjective but would love to get your thoughts.

In terms of notebook, here is an updated version in a ArviZ PR https://github.com/arviz-devs/arviz/pull/510. If you like it I'll merge so we can close the feedback out!

https://github.com/arviz-devs/arviz/blob/8b9528ac046843ffd84f2429caff6fe0ce1c76da/doc/notebooks/XarrayforArviZ.ipynb
