### 2019-01-08 e4b1347
add test and example
