### 2019-01-08 0b14c03
KDE: Display quantiles (#479)
* display quartiles

* use where

* compute quantiles from kde

* remove unused variable

* add test and example
