### 2019-01-08 2c97ef9
Rename s_eff to effective sample size
