+++ arviz/tests/test_diagnostics.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from ..data import load_arviz_data
-from ..stats import rhat, effective_s, geweke
+from ..stats import rhat, effective_sample_size, geweke
 
 GOOD_RHAT = 1.1
 
@@ -33,15 +33,15 @@ class TestDiagnostics:
         r_hat = rhat(np.vstack([20 + np.random.randn(1, 100), np.random.randn(1, 100)]))
         assert 1 / GOOD_RHAT > r_hat or GOOD_RHAT < r_hat
 
-    def test_effective_s_array(self):
-        eff_s = effective_s(np.random.randn(4, 100))
-        assert eff_s > 100
-        assert eff_s < 800
+    def test_effective_sample_size_array(self):
+        eff_n_hat = effective_sample_size(np.random.randn(4, 100))
+        assert eff_n_hat > 100
+        assert eff_n_hat < 800
 
     @pytest.mark.parametrize("var_names", (None, "mu", ["mu", "tau"]))
-    def test_effective_s_dataset(self, data, var_names):
-        eff_s = effective_s(data, var_names=var_names)
-        assert eff_s.mu > 100  # This might break if the data is regenerated
+    def test_effective_sample_size_dataset(self, data, var_names):
+        eff_n_hat = effective_sample_size(data, var_names=var_names)
+        assert eff_n_hat.mu > 100  # This might break if the data is regenerated
 
     def test_geweke(self):
         first = 0.1
