### 2019-01-09 4c56942
Rename N_effective to S_effective (#505)
* Rename n_eff to s_eff

* Rename tests for s_eff

* Rename s_eff to effective sample size

* Put reference all in one line
