### 2019-01-09 b0307f1
add conditional_jit decorator and speed up _fast_kde 2x
