### 2019-01-09 d551b3c
allow to pass kwargs
