### 2019-01-10 b221268
fix lines tests in plot_trace
