### 2019-01-11 72070a4
Increase coverage (#503)
* add tests and fix bugs

* hack dictionary test

* fix lines tests in plot_trace

* Fix func format

* fix
