+++ arviz/tests/test_plots.py
@@ -1,6 +1,5 @@
 # pylint: disable=redefined-outer-name
 import os
-import time
 from unittest.mock import MagicMock
 import matplotlib.pyplot as plt
 from pandas import DataFrame
@@ -55,19 +54,7 @@ def clean_plots(request, save_figs):
 
     def fin():
         if save_figs is not None:
-
-            # Retry save three times
-            for i in range(3):
-                try:
-                    plt.savefig("{0}.png".format(os.path.join(save_figs, request.node.name)))
-
-                except Exception as err:  # pylint: disable=broad-except
-                    if i == 2:
-                        raise err
-                    time.sleep(0.250)
-                else:
-                    break
-
+            plt.savefig("{0}.png".format(os.path.join(save_figs, request.node.name)))
         plt.close("all")
 
     request.addfinalizer(fin)
@@ -114,8 +101,11 @@ def fig_ax():
     [
         {"point_estimate": "mean"},
         {"point_estimate": "median"},
+        {"credible_interval": 0.94},
+        {"credible_interval": 1},
         {"outline": True},
         {"colors": ["g", "b", "r", "y"]},
+        {"colors": "k"},
         {"hpd_markers": ["v"]},
         {"shade": 1},
     ],
@@ -133,6 +123,20 @@ def test_plot_density_discrete(discrete_model):
     assert axes.shape[0] == 2
 
 
+def test_plot_density_bad_kwargs(models):
+    obj = [
+        getattr(models, model_fit) for model_fit in ["pymc3_fit", "stan_fit", "pyro_fit", "tfp_fit"]
+    ]
+    with pytest.raises(ValueError):
+        plot_density(obj, point_estimate="bad_value")
+
+    with pytest.raises(ValueError):
+        plot_density(obj, data_labels=["bad_value_{}".format(i) for i in range(len(obj) + 10)])
+
+    with pytest.raises(ValueError):
+        plot_density(obj, credible_interval=2)
+
+
 @pytest.mark.parametrize("model_fit", ["pymc3_fit", "stan_fit", "pyro_fit", "tfp_fit"])
 @pytest.mark.parametrize(
     "kwargs",
@@ -144,7 +148,7 @@ def test_plot_density_discrete(discrete_model):
         {"divergences": "top"},
         {"divergences": False},
         {"lines": [("mu", {}, [1, 2])]},
-        {"lines": [("mu", 0)]},
+        {"lines": [("mu", {}, 8)]},
     ],
 )
 def test_plot_trace(models, model_fit, kwargs):
@@ -175,6 +179,7 @@ def test_plot_trace_discrete(discrete_model):
         ({"r_hat": True, "quartiles": False}, 2),
         ({"var_names": ["mu"], "colors": "C0", "eff_n": True, "combined": True}, 2),
         ({"kind": "ridgeplot", "r_hat": True, "eff_n": True}, 3),
+        ({"kind": "ridgeplot", "r_hat": True, "eff_n": True, "ridgeplot_alpha": 0}, 3),
     ],
 )
 def test_plot_forest(models, model_fits, args_expected):
@@ -189,11 +194,31 @@ def test_plot_forest_single_value():
     assert axes.shape
 
 
+@pytest.mark.parametrize(
+    "model_fits",
+    [["tfp_fit"], ["pyro_fit"], ["pymc3_fit"], ["stan_fit"], ["pymc3_fit", "stan_fit"]],
+)
+def test_plot_forest_bad(models, model_fits):
[…]
