+++ arviz/tests/test_utils.py
@@ -1,5 +1,5 @@
 """
-Tests for arviz.utils
+Tests for arviz.utils.
 """
 import pytest
 from arviz.utils import _var_names
