### 2019-01-11 b36d9ba
add stable logsumexp
