### 2019-01-11 b750bd5
use b_inv
