### 2019-01-12 4da0893
fix dtype casting and axis is None
