### 2019-01-12 7851f79
fix test back and fix logsumexp missing subtraction
