### 2019-01-12 dbf29be
add dict tests
