+++ arviz/data/__init__.py
@@ -5,6 +5,7 @@ from .datasets import load_arviz_data, list_datasets, clear_data_home
 from .base import numpy_to_data_array, dict_to_dataset
 from .converters import convert_to_dataset, convert_to_inference_data
 from .io_cmdstan import from_cmdstan
+from .io_dict import from_dict
 from .io_pymc3 import from_pymc3
 from .io_pystan import from_pystan
 from .io_emcee import from_emcee
@@ -24,6 +25,7 @@ __all__ = [
     "from_pystan",
     "from_emcee",
     "from_cmdstan",
+    "from_dict",
     "from_pyro",
     "from_tfp",
     "from_netcdf",
