### 2019-01-12 5a6a673
from_python to from_dict
