### 2019-01-12 65d0930
add tests for edge cases
