### 2019-01-12 ddcfe9a
Fix black error for CI
