### 2019-01-13 4697100
add test against scipy
