### 2019-01-13 c875819
support multichains and add jitter
