+++ arviz/plots/ppcplot.py
@@ -1,4 +1,5 @@
 """Posterior predictive plot."""
+from numbers import Integral
 import numpy as np
 from .kdeplot import plot_kde, _fast_kde
 from .plot_utils import (
@@ -14,7 +15,7 @@ from ..utils import _var_names
 def plot_ppc(
     data,
     kind="density",
-    alpha=0.2,
+    alpha=None,
     mean=True,
     figsize=None,
     textsize=None,
@@ -24,6 +25,7 @@ def plot_ppc(
     flatten=None,
     num_pp_samples=None,
     random_seed=None,
+    jitter=0.1,
 ):
     """
     Plot for Posterior Predictive checks.
@@ -36,7 +38,8 @@ def plot_ppc(
     kind : str
         Type of plot to display (density, cumulative, or scatter). Defaults to density.
     alpha : float
-        Opacity of posterior predictive density curves. Defaults to 0.2.
+        Opacity of posterior predictive density curves. Defaults to 0.2 for kind = density
+        and cumulative, for scatter defaults to 0.7
     mean : bool
         Whether or not to plot the mean posterior predictive distribution. Defaults to True
     figsize : tuple
@@ -67,13 +70,17 @@ def plot_ppc(
         specified in the coords argument. Defaults to flattening all of the dimensions.
     num_pp_samples : int
         The number of posterior predictive samples to plot.
-        It defaults to a maximum of 30 samples for `kind` = 'scatter'.
+        It defaults to a maximum of 5 samples for `kind` = 'scatter' and
+        will set jitter to 0.1 unless defined otherwise.
         Otherwise it defaults to all provided samples.
     random_seed : int
         Random number generator seed passed to numpy.random.seed to allow
         reproducibility of the plot. By default, no seed will be provided
         and the plot will change each call if a random sample is specified
         by `num_pp_samples`.
+    jitter : float
+        If kind is "scatter", jitter will add random normal noise to the height
+        of the ppc samples and observed data. By default 0.1.
 
     Returns
     -------
@@ -133,6 +140,12 @@ def plot_ppc(
     if data_pairs is None:
         data_pairs = {}
 
+    if alpha is None:
+        if kind.lower() == "scatter":
+            alpha = 0.7
+        else:
+            alpha = 0.2
+
     observed = data.observed_data
     posterior_predictive = data.posterior_predictive
 
@@ -153,14 +166,14 @@ def plot_ppc(
     total_pp_samples = posterior_predictive.sizes["chain"] * posterior_predictive.sizes["draw"]
     if num_pp_samples is None:
         if kind == "scatter":
-            num_pp_samples = min(30, total_pp_samples)
+            num_pp_samples = min(5, total_pp_samples)
         else:
             num_pp_samples = total_pp_samples
 
     if (
-        not isinstance(num_pp_samples, int)
-        or not num_pp_samples >= 1
-        or not num_pp_samples <= total_pp_samples
+        not isinstance(num_pp_samples, Integral)
+        or num_pp_samples < 1
+        or num_pp_samples > total_pp_samples
     ):
         raise TypeError(
             "`num_pp_samples` must be an integer between 1 and "
@@ -201,9 +214,12 @@ def plot_ppc(
 
         # flatten non-specified dimensions
         obs_vals = obs_vals.flatten()
-        pp_vals = pp_vals.squeeze()
+        if 1 in pp_vals.shape[1:]:
+            pp_vals = pp_vals.squeeze(axis=tuple(a for a, _ in enumerate(pp_vals.shape[1:], 1)))
         if len(pp_vals.shape) > 2:
-            pp_vals = pp_vals.reshape((pp_vals.shape[0], np.prod(pp_vals.shape[1:])))
+            pp_vals = pp_vals.reshape(
[…]
