### 2019-01-13 50fafa7
redefine jitter scale
