### 2019-01-13 76ff708
Add testing for jit decorator
