### 2019-01-13 4d27d39
Fix bug in tests
Fix linting errors
