### 2019-01-13 addd9e2
Fix additional linting error
