### 2019-01-13 cd6a17f
Reword docstring and comment
