+++ arviz/tests/test_utils.py
@@ -29,10 +29,10 @@ def utils_with_numba_import_fail(monkeypatch):
 
 
 def test_utils_fixture(utils_with_numba_import_fail):
-    """Meta test of utils fixture to ensure things are mocked out correctly"""
+    """Test of utils fixture to ensure mock is applied correctly"""
 
-    # If Numba doesn't exist in dev environment this should fail immediately
-    import numba  # pylint: disable=W0611,W0612
+    # If Numba doesn't exist in dev environment this will raise an ImportError
+    import numba  # pylint: disable=unused-import,W0612
 
     with pytest.raises(ImportError):
         utils_with_numba_import_fail.importlib.import_module("numba")
