### 2019-01-13 16342e1
optional legend in kdeplot
