### 2019-01-13 5d6d10b
redefine jitter so limit works as no jitter
