### 2019-01-13 a7de9f2
fix dtype issue
