### 2019-01-13 e88e1bf
From dict function (#524)
* add io_dict

* add dict tests

* add tests for edge cases
