+++ arviz/data/io_dict.py
@@ -1,0 +1,182 @@
+"""Dictionary specific conversion code."""
+import logging
+
+import numpy as np
+import xarray as xr
+
+from .inference_data import InferenceData
+from .base import requires, dict_to_dataset, generate_dims_coords, make_attrs
+
+_log = logging.getLogger("arviz")
+
+
+class DictConverter:
+    """Encapsulate Dictionary specific logic."""
+
+    def __init__(
+        self,
+        *,
+        posterior=None,
+        posterior_predictive=None,
+        sample_stats=None,
+        prior=None,
+        prior_predictive=None,
+        sample_stats_prior=None,
+        observed_data=None,
+        coords=None,
+        dims=None,
+    ):
+        self.posterior = posterior
+        self.posterior_predictive = posterior_predictive
+        self.sample_stats = sample_stats
+        self.prior = prior
+        self.prior_predictive = prior_predictive
+        self.sample_stats_prior = sample_stats_prior
+        self.observed_data = observed_data
+        self.coords = coords
+        self.dims = dims
+
+    @requires("posterior")
+    def posterior_to_xarray(self):
+        """Convert posterior samples to xarray."""
+        data = self.posterior
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.posterior is not a dictionary")
+
+        if "log_likelihood" in data:
+            _log.warning(
+                "log_likelihood found in posterior."
+                " For stats functions log_likelihood needs to be in sample_stats."
+            )
+
+        return dict_to_dataset(data, library=None, coords=self.coords, dims=self.dims)
+
+    @requires("sample_stats")
+    def sample_stats_to_xarray(self):
+        """Convert sample_stats samples to xarray."""
+        data = self.sample_stats
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.sample_stats is not a dictionary")
+
+        return dict_to_dataset(data, library=None, coords=self.coords, dims=self.dims)
+
+    @requires("posterior_predictive")
+    def posterior_predictive_to_xarray(self):
+        """Convert posterior_predictive samples to xarray."""
+        data = self.posterior_predictive
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.posterior_predictive is not a dictionary")
+
+        return dict_to_dataset(data, library=None, coords=self.coords, dims=self.dims)
+
+    @requires("prior")
+    def prior_to_xarray(self):
+        """Convert prior samples to xarray."""
+        data = self.prior
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.prior is not a dictionary")
+
+        return dict_to_dataset(data, library=None, coords=self.coords, dims=self.dims)
+
+    @requires("sample_stats_prior")
+    def sample_stats_prior_to_xarray(self):
+        """Convert sample_stats_prior samples to xarray."""
+        data = self.sample_stats_prior
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.sample_stats_prior is not a dictionary")
+
+        return dict_to_dataset(data, library=None, coords=self.coords, dims=self.dims)
+
+    @requires("prior_predictive")
+    def prior_predictive_to_xarray(self):
+        """Convert prior_predictive samples to xarray."""
+        data = self.prior_predictive
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.prior_predictive is not a dictionary")
+
+        return dict_to_dataset(data, library=None, coords=self.coords, dims=self.dims)
+
+    @requires("observed_data")
+    def observed_data_to_xarray(self):
+        """Convert observed_data to xarray."""
+        data = self.observed_data
+        if not isinstance(data, dict):
+            raise TypeError("DictConverter.observed_data is not a dictionary")
[…]
