### 2019-01-13 e7bee34
add tests for multichain, multidimensional and data_pairs
