### 2019-01-13 de100a7
fix b, b_inv behaviour
