### 2019-01-14 cae2218
test when b_inv=0
