### 2019-01-14 876ea97
np.all
