### 2019-01-14 f5d9e1a
y to y_mean
