### 2019-01-14 06683c4
Fix pymc3 typo
