### 2019-01-14 1cba922
default jitter to None
