### 2019-01-14 263de2c
Fix travis
Removed code was not needed due to docker.
