### 2019-01-14 0b9bb6c
add stable logsumexp (#522)
* add stable logsumexp

* use b_inv

* handle none

* Change comparison

* fix dtype casting and axis is None

* fix test back and fix logsumexp missing subtraction

* more tests

* fix tests

* add test against scipy

* fix b, b_inv behaviour

* test when b_inv=0
