+++ arviz/tests/test_stats.py
@@ -3,11 +3,13 @@ from copy import deepcopy
 import numpy as np
 from numpy.testing import assert_almost_equal, assert_array_almost_equal, assert_array_less
 import pytest
+from scipy.special import logsumexp
 from scipy.stats import linregress
+
 
 from ..data import load_arviz_data
 from ..stats import bfmi, compare, hpd, loo, r2_score, waic, psislw, summary
-from ..stats.stats import _gpinv, _mc_error
+from ..stats.stats import _gpinv, _mc_error, _logsumexp
 
 
 @pytest.fixture(scope="session")
@@ -140,11 +142,6 @@ def test_waic_bad(centered_eight):
         waic(centered_eight)
 
 
-@pytest.mark.xfail(
-    reason="Issue #509. "
-    "Numerical accuracy (logsumexp) prevents function to throw a warning."
-    "See https://github.com/arviz-devs/arviz/issues/509"
-)
 def test_waic_warning(centered_eight):
     centered_eight = deepcopy(centered_eight)
     centered_eight.sample_stats["log_likelihood"][:, :250, 1] = 10
@@ -217,3 +214,64 @@ def test_gpinv(probs, kappa, sigma):
     else:
         probs = np.array([-0.1, 0.1, 0.1, 0.2, 0.3])
     assert len(_gpinv(probs, kappa, sigma)) == len(probs)
+
+
+@pytest.mark.parametrize("ary_dtype", [np.float64, np.float32, np.int32, np.int64])
+@pytest.mark.parametrize("axis", [None, 0, 1, (-2, -1)])
+@pytest.mark.parametrize("b", [None, 0, 1 / 100, 1 / 101])
+@pytest.mark.parametrize("keepdims", [True, False])
+def test_logsumexp_b(ary_dtype, axis, b, keepdims):
+    """Test ArviZ implementation of logsumexp.
+
+    Test also compares against Scipy implementation.
+    Case where b=None, they are equal. (N=len(ary))
+    Second case where b=x, and x is 1/(number of elements), they are almost equal.
+
+    Test tests against b parameter.
+    """
+    np.random.seed(17)
+    ary = np.random.randn(100, 101).astype(ary_dtype)
+    assert _logsumexp(ary=ary, axis=axis, b=b, keepdims=keepdims, copy=True) is not None
+    ary = ary.copy()
+    assert _logsumexp(ary=ary, axis=axis, b=b, keepdims=keepdims, copy=False) is not None
+    out = np.empty(5)
+    assert _logsumexp(ary=np.random.randn(10, 5), axis=0, out=out) is not None
+
+    # Scipy implementation
+    scipy_results = logsumexp(ary, b=b, axis=axis, keepdims=keepdims)
+    arviz_results = _logsumexp(ary, b=b, axis=axis, keepdims=keepdims)
+
+    assert_array_almost_equal(scipy_results, arviz_results)
+
+
+@pytest.mark.parametrize("ary_dtype", [np.float64, np.float32, np.int32, np.int64])
+@pytest.mark.parametrize("axis", [None, 0, 1, (-2, -1)])
+@pytest.mark.parametrize("b_inv", [None, 0, 100, 101])
+@pytest.mark.parametrize("keepdims", [True, False])
+def test_logsumexp_b_inv(ary_dtype, axis, b_inv, keepdims):
+    """Test ArviZ implementation of logsumexp.
+
+    Test also compares against Scipy implementation.
+    Case where b=None, they are equal. (N=len(ary))
+    Second case where b=x, and x is 1/(number of elements), they are almost equal.
+
+    Test tests against b_inv parameter.
+    """
+    np.random.seed(17)
[…]
