+++ arviz/stats/stats.py
@@ -1,9 +1,9 @@
 """Statistical functions in ArviZ."""
 import warnings
+from collections.abc import Sequence
 
 import numpy as np
 import pandas as pd
-from scipy.special import logsumexp
 import scipy.stats as st
 from scipy.optimize import minimize
 import xarray as xr
@@ -281,6 +281,62 @@ def hpd(x, credible_interval=0.94, circular=False):
     return np.array([hdi_min, hdi_max])
 
 
+def _logsumexp(ary, *, b=None, b_inv=None, axis=None, keepdims=False, out=None, copy=True):
+    """Stable logsumexp when b >= 0 and b is scalar.
+
+    b_inv overwrites b unless b_inv is None.
+    """
+    # check dimensions for result arrays
+    ary = np.asarray(ary)
+    if ary.dtype.kind == "i":
+        ary = ary.astype(np.float64)
+    dtype = ary.dtype.type
+    shape = ary.shape
+    shape_len = len(shape)
+    if isinstance(axis, Sequence):
+        axis = tuple(axis_i if axis_i >= 0 else shape_len + axis_i for axis_i in axis)
+        agroup = axis
+    else:
+        axis = axis if (axis is None) or (axis >= 0) else shape_len + axis
+        agroup = (axis,)
+    shape_max = (
+        tuple(1 for _ in shape)
+        if axis is None
+        else tuple(1 if i in agroup else d for i, d in enumerate(shape))
+    )
+    # create result arrays
+    if out is None:
+        if not keepdims:
+            out_shape = (
+                tuple()
+                if axis is None
+                else tuple(d for i, d in enumerate(shape) if i not in agroup)
+            )
+        else:
+            out_shape = shape_max
+        out = np.empty(out_shape, dtype=dtype)
+    if b_inv == 0:
+        return np.full_like(out, np.inf, dtype=dtype) if out.shape else np.inf
+    if b_inv is None and b == 0:
+        return np.full_like(out, -np.inf) if out.shape else -np.inf
+    ary_max = np.empty(shape_max, dtype=dtype)
+    # calculations
+    ary.max(axis=axis, keepdims=True, out=ary_max)
+    if copy:
+        ary = ary.copy()
+    ary -= ary_max
+    np.exp(ary, out=ary)
+    ary.sum(axis=axis, keepdims=keepdims, out=out)
+    np.log(out, out=out)
+    if b_inv is not None:
+        ary_max -= np.log(b_inv)
+    elif b:
+        ary_max += np.log(b)
+    out += ary_max.squeeze() if not keepdims else ary_max
+    # transform to scalar if possible
+    return out if out.shape else dtype(out)
+
+
 def loo(data, pointwise=False, reff=None):
     """Pareto-smoothed importance sampling leave-one-out cross-validation.
 
@@ -346,11 +402,11 @@ def loo(data, pointwise=False, reff=None):
         )
         warn_mg = 1
 
-    loo_lppd_i = -2 * logsumexp(log_weights, axis=0)
+    loo_lppd_i = -2 * _logsumexp(log_weights, axis=0)
     loo_lppd = loo_lppd_i.sum()
     loo_lppd_se = (len(loo_lppd_i) * np.var(loo_lppd_i)) ** 0.5
 
-    lppd = np.sum(logsumexp(log_likelihood, axis=0, b=1.0 / log_likelihood.shape[0]))
+    lppd = np.sum(_logsumexp(log_likelihood, axis=0, b_inv=log_likelihood.shape[0]))
     p_loo = lppd + (0.5 * loo_lppd)
 
     if pointwise:
@@ -432,7 +488,7 @@ def psislw(log_weights, reff=1.0):
[…]
