### 2019-01-14 9844c91
Add conditional_jit else test
