### 2019-01-14 6cfb3d3
Add docstring to conditiona_jit
