+++ arviz/utils.py
@@ -24,7 +24,23 @@ def _var_names(var_names):
 
 
 def conditional_jit(function=None, **kwargs):  # noqa: D202
-    """Use numba's jit decorator if numba is installed."""
+    """Use numba's jit decorator if numba is installed.
+
+    Notes
+    -----
+        If called without arguments  then return wrapped function.
+
+        @conditional_jit
+        def my_func():
+            return
+
+        else called with arguments
+
+        @@conditional_jit(nopython=True)
+        def my_func():
+            return
+
+    """
 
     def wrapper(function):
         try:
@@ -34,6 +50,7 @@ def conditional_jit(function=None, **kwargs):  # noqa: D202
         except ImportError:
             return function
 
+    #
     if function:
         return wrapper(function)
     else:
