### 2019-01-14 6614218
Fix linting errors
