### 2019-01-14 97812d8
Fix typo in conditional_jit
