### 2019-01-14 a8dc99a
Remove extra space in docstring
