+++ arviz/tests/test_utils.py
@@ -68,7 +68,7 @@ def test_conditional_jit_numba_decorator():
 
 
 def test_conditional_jit_numba_decorator_keyword(monkeypatch):
-    """Checks else statement and JIT keyword argument """
+    """Checks else statement and JIT keyword argument"""
     from arviz import utils
 
     # Mock import lib to return numba with hit method which returns a function that returns kwargs
