### 2019-01-15 ea62b11
add conditional_jit decorator and speed up _fast_kde 2x (#515)
* add conditional_jit decorator and speed up _fast_kde 2x

* allow to pass kwargs

* remove blank space

* Fix black error for CI

* Add testing for jit decorator

* Fix bug in tests

Fix linting errors

* Fix additional linting error

* Reword docstring and comment

* Add conditional_jit else test

* Add docstring to conditiona_jit

* Fix linting errors

* Fix typo in conditional_jit

* Remove extra space in docstring
