### 2019-01-15 75a2746
Add additional tests for plot_forest
