### 2019-01-15 836d9b9
Added rope to forestplot.py in forestplot() (#448)
Added rope to forestplot.py in forestplot()
