### 2019-01-15 dcc1636
Support multichain ppc in plot_ppc (#526)
* support multichains and add jitter

* redefine jitter scale

* reshape

* fix stuff

* optional legend in kdeplot

* redefine jitter so limit works as no jitter

* fix dtype issue

* add tests for multichain, multidimensional and data_pairs

* add tests

* fix flatten bug for pp with data_pairs

* np.all

* flatten_pp defaults

* default jitter to None
