+++ arviz/plots/ppcplot.py
@@ -1,4 +1,5 @@
 """Posterior predictive plot."""
+from numbers import Integral
 import numpy as np
 from .kdeplot import plot_kde, _fast_kde
 from .plot_utils import (
@@ -14,7 +15,7 @@ from ..utils import _var_names
 def plot_ppc(
     data,
     kind="density",
-    alpha=0.2,
+    alpha=None,
     mean=True,
     figsize=None,
     textsize=None,
@@ -22,8 +23,11 @@ def plot_ppc(
     var_names=None,
     coords=None,
     flatten=None,
+    flatten_pp=None,
     num_pp_samples=None,
     random_seed=None,
+    jitter=None,
+    legend=True,
 ):
     """
     Plot for Posterior Predictive checks.
@@ -36,7 +40,8 @@ def plot_ppc(
     kind : str
         Type of plot to display (density, cumulative, or scatter). Defaults to density.
     alpha : float
-        Opacity of posterior predictive density curves. Defaults to 0.2.
+        Opacity of posterior predictive density curves. Defaults to 0.2 for kind = density
+        and cumulative, for scatter defaults to 0.7
     mean : bool
         Whether or not to plot the mean posterior predictive distribution. Defaults to True
     figsize : tuple
@@ -63,17 +68,28 @@ def plot_ppc(
         that dimension. Defaults to including all coordinates for all
         dimensions if None.
     flatten : list
-        List of dimensions to flatten. Only flattens across the coordinates
+        List of dimensions to flatten in observed_data. Only flattens across the coordinates
         specified in the coords argument. Defaults to flattening all of the dimensions.
+    flatten_pp : list
+        List of dimensions to flatten in posterior_predictive. Only flattens across the coordinates
+        specified in the coords argument. Defaults to flattening all of the dimensions.
+        Dimensions should match flatten excluding dimensions for data_pairs parameters.
+        If flatten is defined and flatten_pp is None, then `flatten_pp=flatten`.
     num_pp_samples : int
         The number of posterior predictive samples to plot.
-        It defaults to a maximum of 30 samples for `kind` = 'scatter'.
+        It defaults to a maximum of 5 samples for `kind` = 'scatter' and
+        will set jitter to 0.7 unless defined otherwise.
         Otherwise it defaults to all provided samples.
     random_seed : int
         Random number generator seed passed to numpy.random.seed to allow
         reproducibility of the plot. By default, no seed will be provided
         and the plot will change each call if a random sample is specified
         by `num_pp_samples`.
+    jitter : float
+        If kind is "scatter", jitter will add random uniform noise to the height
+        of the ppc samples and observed data. By default 0.
+    legend : bool
+        Add legend to figure. By default True.
 
     Returns
     -------
@@ -133,6 +149,16 @@ def plot_ppc(
     if data_pairs is None:
         data_pairs = {}
 
+    if alpha is None:
+        if kind.lower() == "scatter":
+            alpha = 0.7
+        else:
+            alpha = 0.2
+
+    if jitter is None:
+        jitter = 0.0
+    assert jitter >= 0.0
+
     observed = data.observed_data
     posterior_predictive = data.posterior_predictive
 
@@ -141,6 +167,10 @@ def plot_ppc(
     var_names = _var_names(var_names)
     pp_var_names = [data_pairs.get(var, var) for var in var_names]
 
+    if flatten_pp is None and flatten is None:
+        flatten_pp = list(posterior_predictive.dims.keys())
+    elif flatten_pp is None:
+        flatten_pp = flatten
     if flatten is None:
         flatten = list(observed.dims.keys())
 
@@ -153,14 +183,14 @@ def plot_ppc(
[…]
