+++ arviz/tests/test_plots.py
@@ -9,8 +9,8 @@ import pytest
 import pymc3 as pm
 
 
-from ..data import from_dict, from_pymc3, InferenceData
-from ..stats import compare
+from ..data import from_pymc3, InferenceData
+from ..stats import compare, psislw
 from .helpers import eight_schools_params, load_cached_models  # pylint: disable=unused-import
 from ..plots import (
     plot_density,
@@ -29,8 +29,7 @@ from ..plots import (
     plot_khat,
     plot_hpd,
 )
-
-from ..stats import psislw
+from ..plots.kdeplot import _histogram
 
 np.random.seed(0)
 
@@ -135,6 +134,14 @@ def test_plot_density_bad_kwargs(models):
 
     with pytest.raises(ValueError):
         plot_density(obj, credible_interval=2)
+
+
+def test_histogram_without_numba():
+    random_state = np.random.RandomState(seed=17)
+    arr = random_state.randn(100)
+    numpy_bins, _ = np.histogram(arr, bins=10)
+    hist_bins = _histogram(arr, n_bins=10)
+    assert np.all(numpy_bins == hist_bins)
 
 
 @pytest.mark.parametrize("model_fit", ["pymc3_fit", "stan_fit", "pyro_fit", "tfp_fit"])
@@ -365,26 +372,10 @@ def test_plot_pair_bad(models, model_fit):
 
 
 @pytest.mark.parametrize("kind", ["density", "cumulative", "scatter"])
-@pytest.mark.parametrize("alpha", [None, 0.2, 1])
-def test_plot_ppc(models, pymc3_sample_ppc, kind, alpha):
+def test_plot_ppc(models, pymc3_sample_ppc, kind):
     data = from_pymc3(trace=models.pymc3_fit, posterior_predictive=pymc3_sample_ppc)
-    axes = plot_ppc(data, kind=kind, alpha=alpha, random_seed=3)
+    axes = plot_ppc(data, kind=kind, random_seed=3)
     assert axes
-
-
-@pytest.mark.parametrize("kind", ["density", "cumulative", "scatter"])
-@pytest.mark.parametrize("jitter", [None, 0, 0.1, 1, 3])
-def test_plot_ppc_multichain(kind, jitter):
-    np.random.seed(23)
-    data = from_dict(
-        posterior_predictive={
-            "x": np.random.randn(4, 100, 30),
-            "y_hat": np.random.randn(4, 100, 3, 10),
-        },
-        observed_data={"x": np.random.randn(30), "y": np.random.randn(3, 10)},
-    )
-    axes = plot_ppc(data, kind=kind, data_pairs={"y": "y_hat"}, jitter=jitter, random_seed=3)
-    assert np.all(axes)
 
 
 @pytest.mark.parametrize("kind", ["density", "cumulative", "scatter"])
