### 2019-01-15 5978085
RandomState no-member pylint
