### 2019-01-15 ae69b67
NUMBA_DISABLE_JIT=1 for .travis file
