### 2019-01-15 db73ac5
test that placeholder func returns None
