### 2019-01-15 243e191
Remove redundant docker line in travisyml (#531)
