### 2019-01-15 49844be
rewrite custom jit
