### 2019-01-15 bfeee22
remove unneeded pylint disable
