### 2019-01-15 98b446e
Fix pylint
