### 2019-01-15 4e5a032
Run tests without numba (#534)
* NUMBA_DISABLE_JIT=1 for .travis file

* import order

* test that placeholder func returns None

* rewrite custom jit

* better jit

* remove unneeded pylint disable

* Fix pylint
