### 2019-01-15 arviz-devs/arviz#535 @canyon289
Include Zenodo and JOSS Badges in README and docs
## Tell us about it

Add badge labels for Zenodo and JOSS to README and docs

https://zenodo.org/record/2540945#.XD5DP9-YXmE
https://github.com/openjournals/joss-reviews/issues/1143#issuecomment-454512886

## Thoughts on implementation

Code and be copied and pasted from both sources
