### 2019-01-16 arviz-devs/arviz#535 @canyon289
Closed by https://github.com/arviz-devs/arviz/pull/537
