### 2019-01-16 arviz-devs/arviz#535 @canyon289
Still need them on docs. Reopening.

I can do this unless someone beats me to it
