### 2019-01-23 arviz-devs/arviz#535 @ahartikainen
Fixed by #543
